+++ changelog.md
@@ -6,6 +6,7 @@
 - Added support for dev-legends
 
 ### Other
+- When an image file already exists in the game files during import the user is now asked if it should be overwritten or if the import should be canceled.
 - When a new update is available and the GitHub repository is opened, the specific release page is now opened
 - Optimized localisation files
 - Error message is now written to the console when the import fails because image files could not be copied